% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_independence.R
\name{monthly_investments_till_financial_independence}
\alias{monthly_investments_till_financial_independence}
\title{Monthly investment for Financial Independence}
\usage{
monthly_investments_till_financial_independence(
  current_monthly_expense,
  start_year,
  withdrawal_period,
  currency_inflation,
  lifestyle_inflation,
  mf_growth,
  sip_growth,
  life_expectancy,
  age_current
)
}
\description{
Get monthly investment needed into mutual funds as of now for achieving fin independence by a given period of time.
Returns the monthly investment needed for the current year. And if sip_growth > 0, this means the monthly investment is planned to increase by the same every year (maybe due to more earning, inflation etc)
}
