% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean.R
\name{read_and_clean_raw_data}
\alias{read_and_clean_raw_data}
\title{Read and clean raw Monefy exported csv}
\usage{
read_and_clean_raw_data(
  file,
  delimiter_character = ",",
  decimal_separator = ".",
  write_path = "expenses_clean.rds"
)
}
\arguments{
\item{file}{path to the raw csv exported from monefy app}

\item{decimal_separator}{decimal separator chosen during export}

\item{write_path}{path to write the cleaned rds file into}

\item{delimer_character}{delimiter character chosen during export}
}
\description{
Read and clean raw Monefy exported csv

Read, clean and write Monefy exported csv data to an rds file
}
